se actualizo

+++ src/main/java/com/aceruservicios/repository/RankingRepo.java
@@ -3,6 +3,6 @@ package com.aceruservicios.repository;
 import com.aceruservicios.entity.Ranking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RankingRepo extends JpaRepository<Ranking, Integer> {
+public interface RankingRepo extends JpaRepository<Ranking, Long> {
 
 }
